Crash-bug fix, minor fixes, fuel balancing
I fixed a crash-bug that was causing a race-condition segfault if the
mech fuel window was left open when beaming up/down. During beaming, if
the update function ran, there was a change that player.id() would
return 0, causing world.sendEntityMessage to segfault.

Another change I made to mechfuel.lua was to centralize all of the info
about fuels into a table defined in the init function to avoid
repetition.  The current breakdown I've been experimenting with is
oil=1, liquid erchius=1.8, solid erchius=3.6, unrefinedmechfuel=4, and
mechfuel=8. I realize that 8:1 seems a little over-powered, but I was
experimenting with setting fuel to 0 if the mech explodes to offset
this, when I discovered and fixed the crash-bug described above. If you
want, you can modify the table to return the ratios to what they were
before.

Another smaller bugfix in mechfuel.lua was that, in the old code, if you
had 799/800 fuel and 1000 oil in the fuel slot, when you clicked fuel
your fuel rose to 800/800 fuel, but your oil went to 499. I simplified
the logic around refueling to fix this, although the bug symptons have
been masked by the ratio changes I described above.

One final change included in the above mechfuel.lua changes the names
'Mech fuel' and 'Unrefined' to be 'Mech Fuel' and 'Unrefined Fuel'
respectively, to make them more consistent with other uses of those
names.

I fixed a small bug in catalystfuelrefinery that, if you had a stack of
999 mech fuel, the next refining would reduce your unrefined input by
10, but would produce only 1 new mech fuel. I fixed it so that refining
stops if your stack of mech fuel is >990.

I changed the logic in mechfuelrefinery to mirror that of
catalystfuelrefinery for consistency.

Finally, there were a mix of tabs and spaces in the source files that
were making indentation weird. I fixed it by switching to all spaces.